Changed the way the scrolling was calculated. Instead of calculating it relatively to the middle of the screen, calculte it relatively to the borders. This solution is more clear (IMO) and elegant because; there isn't that ugly code anymore: scroll_x -= screen.w / 2; (...) scroll_x += screen.w / 2;
Besides from that, in case we change the resolution, this makes the playbility to be not affected.

Note: if you think the scrolling is made when Tux is too much in the right or in the left, you can easily change this, by changine the SPACE_X value in world.cpp.