Changed the position of the message.
Marek, I find it confusing, as a player, to have two messages when going to the first level of each "island", so my suggestion is doing like this; put the messages in a fork or something before the level.
Have a look at this example (just go to the abednego area) and check if you also prefer this approach.