Added support for one way roads, as requested by Marek. Possible of usage: (one-way "north-south"), (one-way "south-north"), (one-way "east-west") or (one-way "west-east"). Use only one. And none for both ways.
Worldmap stuff is getting too many expections, which is not bad, but is getting bloated. After the release of 0.1.2, I will port all these new aditions, but in a way to avoid both worldmap dta and source code getting bloated.
The use of strings in worldmap data (like in this case), and arrays of booleans or peraphs bitfields in the code sound like a good choice.